Version 1.9.0 alexa-client-sdk
Changes in this update:

Enhancements

    Added Android SDK support, which includes new implementations of the MediaPlayer, audio recorder, and logger.
    Added the InteractionModel interface, which enables Alexa Routines.
    Optional configuration changes have been introduced. Now a network interface can be specified to connect to the SDK via curl.
    Build options can be configured to support Android.
    Added GUI 1.1 support. The PlaybackController has been extended to support new control functionality, and the System interface has been updated to support SoftwareInfo.

Bug Fixes

    Installation script execution time has been reduced. Now a single branch clone is used, such as the master branch.
    Issue 846 - Fixed a bug where audio stuttered on slow network connections.
    Removed the SpeakerManager constructor check for non-zero speakers.
    Issue 891 - Resolved incorrect offset in the PlaybackFinished event.
    Issue 727 - Fixed an issue where the sample app behaved erratically upon network disconnection/reconnection.
    Issue 910 - Fixed a GCC 8+ compilation issue. Note: issues related to -Wclass-memaccess will still trigger warnings, but won't fail compilation.
    Issue 871 Issue 880 - Fixed compiler warnings.
    Fixed a bug where Ted Talks would not stream via TuneIn.
    Fixed an issue where the PryonLiteKeywordDetector would not restart.
    Fixed an issue where PlaybackStutterStarted and PlaybackStutterFinished events were not being sent due to a missing Gstreamer queue element.
    Fixed a bug where the CapabilitiesDelegate database was not being cleared upon logout.
    Fixed in issue that caused the following compiler warning “class has virtual functions but non-virtual destructor”.
    Fixed a bug where BlueZDeviceManager was not properly destroyed.
    Fixed a bug that occurred when the initializer list was converted to std::unordered_set.
    Fixed a build error that occurred when building with BUILD_TESTING=Off.

Known Issues
* The ACL may encounter issues if audio attachments are received but not consumed.
* SpeechSynthesizerState currently uses GAINING_FOCUS and LOSING_FOCUS as a workaround for handling intermediate state. These states may be removed in a future release.
* The Alexa app doesn't always indicate when a device is successfully connected via Bluetooth.
* Connecting a product to streaming media via Bluetooth will sometimes stop media playback within the source application. Resuming playback through the source application or toggling next/previous will correct playback.
* When a source device is streaming silence via Bluetooth, the Alexa companion app indicates that audio content is streaming.
* The Bluetooth agent assumes that the Bluetooth adapter is always connected to a power source. Disconnecting from a power source during operation is not yet supported.
* On some products, interrupted Bluetooth playback may not resume if other content is locally streamed.
* On Raspberry Pi, when streaming audio via Bluetooth, sometimes the audio stream stutters.
* These CapabilitiesDelegateTest tests have been temporarily disabled to prevent build errors for the Android build: CapabilitiesDelegateTest.withCapabilitiesHappyCase, CapabilitiesDelegateTest.republish, CapabilitiesDelegateTest.testClearData.
* make integration is currently not available for Android. In order to run integration tests on Android, you'll need to manually upload the test binary file along with the any input file. At that point, the adb can be used to run the integration tests.
* On Raspberry Pi running Android Things with HDMI output audio, beginning of speech is truncated when Alexa responds to user TTS.
* Spotify does not completely shut down when the sample app quits.

+++ KWD/KittAi/src/KittAiKeyWordDetector.cpp
@@ -119,7 +119,8 @@ KittAiKeyWordDetector::KittAiKeyWordDetector(
     std::chrono::milliseconds msToPushPerIteration) :
         AbstractKeywordDetector(keyWordObservers, keyWordDetectorStateObservers),
         m_stream{stream},
-        m_maxSamplesPerPush{(audioFormat.sampleRateHz / HERTZ_PER_KILOHERTZ) * msToPushPerIteration.count()} {
+        m_maxSamplesPerPush{
+            static_cast<size_t>((audioFormat.sampleRateHz / HERTZ_PER_KILOHERTZ) * msToPushPerIteration.count())} {
     std::stringstream sensitivities;
     std::stringstream modelPaths;
     for (unsigned int i = 0; i < kittAiConfigurations.size(); ++i) {
